% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFDR.R
\name{ccFDR}
\alias{ccFDR}
\title{Conjunctional conditional false discovery rate}
\usage{
ccFDR(data, p1, p2, p_threshold = 0.001)
}
\arguments{
\item{data}{data frame or (numeric) matrix with p-values as columns}

\item{p1}{p-values for which the ccFDR will be estimated; either the name/index of the columns in
\code{data}, or a vectors of p-values}

\item{p2}{the p-values on which the ccFDR will be conditioned on; either name or vector of p-values
\code{data}, or a vectors of p-values}

\item{p_threshold}{cutoff for pre-filtering the p-values; a vector of either length one (in which
case both sets of pvalues have the same threshold) or length two (where the thresholds refer
to \code{p1} and \code{p2} in that order).}
}
\value{
A data frame with columns cFDR1, cFDR2, and ccFDR: if \code{data} was specified,
the columns are simply added at the end; if only \code{p1} and \code{p2} were specified, a
data frame with five columns (the original p-values, plus the cFDRs and the ccFDR).
}
\description{
Estimate the conjunctional conditional false discovery rate for two set of p-values,
conditional on each other, as measure for pleiotropy.
}
\examples{
data(psynth)
res = ccFDR(psynth, "p1", "p2", p_threshold = 1E-5)
head(res)
head(subset(res, ccFDR < 0.01))
}
