% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFDR.R
\name{cFDR}
\alias{cFDR}
\title{Conditional false discovery rate}
\usage{
cFDR(data, p1, p2, p2_threshold = 0.001)
}
\arguments{
\item{data}{data frame or (numeric) matrix with p-values as columns}

\item{p1}{p-values for which the cFDR will be estimated; either the name of the columns in
\code{data}, or a vectors of p-values}

\item{p2}{the p-values on which the cFDR will be conditioned on; either name or vector of p-values
\code{data}, or a vectors of p-values}

\item{p2_threshold}{cutoff for pre-filtering on \code{p2}: only variants
with \code{p2} \leq \code{p2_threshold} are included}
}
\value{
A data frame consisting of the p-values and the cFDR
}
\description{
Estimate the conditional false discovery rate for a set of p-values,
conditional on a second set of p-values
}
