% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFDR.R
\docType{data}
\name{psynth}
\alias{psynth}
\title{Synthetic SNP p-values}
\format{
A data frame with 1,750,000 rows and 3 variables:
\describe{
\item{SNP}{ids of fictitious SNPs}
\item{p1}{p-values for first phenotype}
\item{p2}{p-values for second phenotype}
}
}
\usage{
psynth
}
\description{
Synthetic p-values for two phenotypes across a moderately larger number of
(fake) SNPs.
}
\keyword{datasets}
